optimise  log statements and unit tests
