+++ src/main/java/uk/gov/hmcts/opal/scheduler/aspect/LogExecutionTime.java
@@ -7,5 +7,5 @@ import java.lang.annotation.Target;
 
 @Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface LogJobExecutionTime {
+public @interface LogExecutionTime {
 }
