+++ src/main/java/uk/gov/hmcts/opal/scheduler/aspect/LogExecutionTimeAspect.java
@@ -4,22 +4,15 @@ import lombok.extern.slf4j.Slf4j;
 import org.aspectj.lang.ProceedingJoinPoint;
 import org.aspectj.lang.annotation.Around;
 import org.aspectj.lang.annotation.Aspect;
-import org.quartz.JobExecutionContext;
 import org.springframework.stereotype.Component;
-
-import java.util.concurrent.TimeUnit;
 
 @Aspect
 @Component
 @Slf4j
-public class LogJobExecutionTimeAspect {
+public class LogExecutionTimeAspect {
 
-    @Around("@annotation(LogJobExecutionTime)")
+    @Around("@annotation(uk.gov.hmcts.opal.scheduler.aspect.LogExecutionTime)")
     public Object logExecutionTime(ProceedingJoinPoint joinPoint) throws Throwable {
-        JobExecutionContext context = (JobExecutionContext) joinPoint.getArgs()[0];
-        String jobName = context.getJobDetail().getKey().getName();
-
-        log.info("Job '{}' started as per fire time {}", jobName, context.getFireTime());
 
         long startTime = System.nanoTime();
         Object result = joinPoint.proceed();
@@ -27,9 +20,7 @@ public class LogJobExecutionTimeAspect {
 
         long executionTime = endTime - startTime;
 
-        log.info("Job '{}' executed in {} ms, next fire time: {}",
-                 jobName, TimeUnit.NANOSECONDS.toMillis(executionTime), context.getNextFireTime()
-        );
+        log.info(joinPoint.getSignature() + " executed in " + executionTime + "ms");
 
         return result;
     }
