+++ src/main/java/uk/gov/hmcts/opal/scheduler/job/LogRetentionJob.java
@@ -7,7 +7,6 @@ import org.quartz.JobExecutionContext;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-import uk.gov.hmcts.opal.scheduler.aspect.LogJobExecutionTime;
 import uk.gov.hmcts.opal.scheduler.model.CronJob;
 import uk.gov.hmcts.opal.scheduler.service.LogRetentionService;
 
@@ -27,7 +26,6 @@ public class LogRetentionJob implements CronJob {
         this.logRetentionService = logRetentionService;
     }
 
-    @LogJobExecutionTime
     @Override
     public void execute(JobExecutionContext context) {
         try {
