+++ src/main/java/uk/gov/hmcts/opal/scheduler/service/AllPayArchiveService.java
@@ -3,6 +3,7 @@ package uk.gov.hmcts.opal.scheduler.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import uk.gov.hmcts.opal.scheduler.aspect.LogExecutionTime;
 import uk.gov.hmcts.opal.sftp.SftpOutboundService;
 
 import java.io.InputStream;
@@ -16,6 +17,7 @@ public class AllPayArchiveService {
 
     private final SftpOutboundService sftpOutboundService;
 
+    @LogExecutionTime
     public void process(String fileName) {
         sftpOutboundService.downloadFile(ALL_PAY_ARCHIVE.getPath(), fileName, this::processFile);
     }
