+++ src/main/java/uk/gov/hmcts/opal/scheduler/service/AutoCheckService.java
@@ -3,6 +3,7 @@ package uk.gov.hmcts.opal.scheduler.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import uk.gov.hmcts.opal.scheduler.aspect.LogExecutionTime;
 import uk.gov.hmcts.opal.sftp.SftpInboundService;
 
 import java.io.InputStream;
@@ -16,6 +17,7 @@ public class AutoCheckService {
 
     private final SftpInboundService sftpInboundService;
 
+    @LogExecutionTime
     public void process(String fileName) {
         sftpInboundService.downloadFile(AUTO_CHEQUES.getPath(), fileName, this::processFile);
     }
